+++ src/components/3d/Scene.tsx
@@ -4,11 +4,16 @@ import Flights from "./Flights.tsx";
 import FlightTrail from "./FlightTrail.tsx";
 import {Suspense} from "react";
 import {EARTH_RADIUS} from "../../constants.ts";
+import {Bloom, EffectComposer, SMAA, Vignette} from "@react-three/postprocessing";
+import {useRecoilValue} from "recoil";
+import {graphicOptionsState} from "../../atoms.ts";
+import {MobileEarth} from "./MobileEarth.tsx";
+import {CountryBorders} from "./countryBorders.tsx";
 import {Earth} from "./Earth.tsx";
-import {Bloom,EffectComposer, Vignette} from "@react-three/postprocessing";
 
 
 function Scene() {
+    const graphicOptions = useRecoilValue(graphicOptionsState)
 
     return (
         <Suspense fallback={
@@ -19,26 +24,61 @@ function Scene() {
             </>
 
         }>
-            <Stars
-                radius={100}
-                depth={50}
-                count={1000}
-                factor={4}
-                saturation={0}
-                fade={true}
+            {
+                graphicOptions.stars ?
+                    <Stars
+                        radius={100}
+                        depth={50}
+                        count={1000}
+                        factor={4}
+                        saturation={0}
+                        fade={true}
 
+                    />
+                    : <></>
+            }
+            <CameraControls
+                minDistance={EARTH_RADIUS + 0.2}
             />
-            <CameraControls/>
-            <Earth/>
+            {
+                graphicOptions.highResolutionEarth ?
+                    <Earth/>
+                    : <MobileEarth/>
+            }
             <pointLight position={[EARTH_RADIUS + 102, 0, 0]}
                         castShadow={true}
             />
             <Flights/>
             <FlightTrail/>
-            <EffectComposer>
-                <Bloom luminanceThreshold={0} luminanceSmoothing={0.9} height={300} />
-                <Vignette eskil={false} offset={0.1} darkness={0.9} />
-            </EffectComposer>
+
+            {
+                graphicOptions.countryBorders ?
+                    <CountryBorders/>
+                    : <></>
+            }
+
+
+            {
+                (graphicOptions.bloom || graphicOptions.vignette || graphicOptions.SMAA) ?
+                    <EffectComposer>
+                        {
+                            graphicOptions.bloom ?
+                                <Bloom luminanceThreshold={0} luminanceSmoothing={0.9} height={300}
+                                /> : <></>
+                        }
+                        {
+                            graphicOptions.vignette ?
+                                <Vignette eskil={false} offset={0.1} darkness={0.9}
+                                /> : <></>
+                        }
+                        {
+                            graphicOptions.SMAA ?
+                                <SMAA/> : <></>
+                        }
+
+                    </EffectComposer>
+                    : <></>
+            }
         </Suspense>
     )
 }